feat: 优化列表刷新并拆分大型模块
