+++ src-tauri/src/database/repos/mod.rs
@@ -1,0 +1,7 @@
+pub mod settings;
+pub mod sync_dto;
+pub mod tag;
+
+pub use settings::SettingsRepository;
+pub use sync_dto::{TagAssocSyncEntry, TagSyncEntry, TagsSyncData};
+pub use tag::{Tag, TagRepository};
