+++ src-tauri/src/webdav/mod.rs
@@ -22,6 +22,9 @@ pub struct WebDavConfig {
     /// 自定义代理地址，如 `http://127.0.0.1:7890` 或 `socks5://127.0.0.1:1080`
     #[serde(default)]
     pub proxy_url: String,
+    /// 是否接受无效 TLS 证书。默认关闭，仅用于自签名 WebDAV 服务兼容。
+    #[serde(default)]
+    pub accept_invalid_certs: bool,
 }
 
 /// 同步选项
@@ -57,6 +60,7 @@ impl Default for SyncOptions {
 
 
 const SYNC_FILENAME: &str = "clipboard_sync.zip";
+const MAX_SYNC_DOWNLOAD_BYTES: u64 = 64 * 1024 * 1024;
 
 /// 媒体文件映射条目（记录每个文件的 hash 和本地路径，用于下载时定位）
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
@@ -89,8 +93,21 @@ pub fn get_or_create_device_id(db: &crate::database::Database) -> String {
 }
 
 /// 计算文件内容的 blake3 hash（hex）
-fn file_hash(data: &[u8]) -> String {
-    blake3::hash(data).to_hex().to_string()
+fn file_hash_from_path(path: &Path) -> Result<(String, u64), String> {
+    let mut file = std::fs::File::open(path)
+        .map_err(|e| format!("读取文件失败 {}: {}", path.display(), e))?;
+    let mut hasher = blake3::Hasher::new();
+    let bytes = std::io::copy(&mut file, &mut hasher)
+        .map_err(|e| format!("计算文件 hash 失败 {}: {}", path.display(), e))?;
+    Ok((hasher.finalize().to_hex().to_string(), bytes))
+}
+
+fn file_len_if_within_limit(path: &Path, max_bytes: i64) -> Option<u64> {
+    let len = std::fs::metadata(path).ok()?.len();
+    if max_bytes >= 0 && len > max_bytes as u64 {
+        return None;
+    }
+    Some(len)
 }
 
 /// 构建 Basic Auth 头
@@ -117,10 +134,14 @@ fn default_proxy_mode() -> String { "system".to_string() }
 /// - none: 不使用任何代理
 /// - custom: 使用自定义代理地址（支持 http/https/socks5）
 fn build_client(config: &WebDavConfig) -> Result<reqwest::blocking::Client, String> {
-    let builder = reqwest::blocking::Client::builder()
+    let mut builder = reqwest::blocking::Client::builder()
         .connect_timeout(std::time::Duration::from_secs(15))
-        .timeout(std::time::Duration::from_secs(60))
-        .danger_accept_invalid_certs(true);
+        .timeout(std::time::Duration::from_secs(60));
+
+    if config.accept_invalid_certs {
+        builder = builder.danger_accept_invalid_certs(true);
+        info!("WebDAV TLS 证书校验已关闭，仅用于自签名服务兼容");
+    }
 
     let builder = crate::proxy::apply_proxy(builder, &config.proxy_mode, &config.proxy_url)?;
 
@@ -338,9 +359,8 @@ pub fn build_media_map(
                     } else {
                         images_dir.join(img_path)
                     };
-                    if let Ok(data) = std::fs::read(&full_path) {
-                        if (data.len() as i64) <= max_image_bytes {
-                            let hash = file_hash(&data);
+                    if file_len_if_within_limit(&full_path, max_image_bytes).is_some()
+                        && let Ok((hash, _)) = file_hash_from_path(&full_path) {
                             if seen_hashes.insert(hash.clone()) {
                                 let ext = full_path.extension()
                                     .unwrap_or_default().to_string_lossy().to_string();
@@ -352,7 +372,6 @@ pub fn build_media_map(
                                     device_id: device_id.to_string(),
                                 });
                             }
-                        }
                     }
                 }
             }
@@ -380,9 +399,8 @@ pub fn build_media_map(
 
                 for file_path in &paths {
                     let p = Path::new(file_path);
-                    if let Ok(data) = std::fs::read(p) {
-                        if (data.len() as i64) <= limit {
-                            let hash = file_hash(&data);
+                    if file_len_if_within_limit(p, limit).is_some()
+                        && let Ok((hash, _)) = file_hash_from_path(p) {
                             if seen_hashes.insert(hash.clone()) {
                                 let ext = p.extension()
                                     .unwrap_or_default().to_string_lossy().to_string();
@@ -394,7 +412,6 @@ pub fn build_media_map(
                                     device_id: device_id.to_string(),
                                 });
                             }
-                        }
                     }
                 }
             }
@@ -440,7 +457,6 @@ pub fn upload_media_files(
             continue;
         }
 
-        // 读取本地文件
         let local_path = if entry.media_type == "image" {
             if Path::new(&entry.local_path).is_absolute() {
                 PathBuf::from(&entry.local_path)
@@ -451,18 +467,31 @@ pub fn upload_media_files(
             PathBuf::from(&entry.local_path)
         };
 
-        if let Ok(data) = std::fs::read(&local_path) {
-            let resp = client.put(&remote_url)
-                .header("Authorization", &auth)
-                .header("Content-Type", "application/octet-stream")
-                .body(data.to_vec())
-                .send()
-                .map_err(|e| format!("上传 {} 失败: {}", remote_path, e))?;
-
-            if resp.status().is_success() {
-                total_bytes += data.len() as u64;
-                uploaded += 1;
+        let file = match std::fs::File::open(&local_path) {
+            Ok(file) => file,
+            Err(e) => {
+                debug!("媒体上传跳过，无法打开 {}: {}", local_path.display(), e);
+                continue;
             }
+        };
+        let data_len = match file.metadata() {
+            Ok(metadata) => metadata.len(),
+            Err(e) => {
+                debug!("媒体上传跳过，无法读取 metadata {}: {}", local_path.display(), e);
+                continue;
+            }
+        };
+        let resp = client.put(&remote_url)
+            .header("Authorization", &auth)
+            .header("Content-Type", "application/octet-stream")
+            .header("Content-Length", data_len)
+            .body(reqwest::blocking::Body::new(file))
+            .send()
+            .map_err(|e| format!("上传 {} 失败: {}", remote_path, e))?;
+
+        if resp.status().is_success() {
+            total_bytes += data_len;
+            uploaded += 1;
         }
     }
 
@@ -512,14 +541,30 @@ pub fn download_missing_media(
             .map_err(|e| format!("下载 {} 失败: {}", remote_path, e))?;
 
         if resp.status().is_success() {
-            let bytes = resp.bytes().map_err(|e| e.to_string())?;
             // 确保父目录存在
             if let Some(parent) = local_path.parent() {
                 let _ = std::fs::create_dir_all(parent);
             }
-            if std::fs::write(&local_path, &bytes).is_ok() {
-                downloaded += 1;
-                info!("媒体下载: {} -> {}", remote_path, local_path.display());
+            let tmp_path = local_path.with_extension(format!(
+                "{}.download",
+                local_path
+                    .extension()
+                    .and_then(|s| s.to_str())
+                    .unwrap_or("tmp")
+            ));
+            let mut body = resp;
+            match std::fs::File::create(&tmp_path)
+                .and_then(|mut file| std::io::copy(&mut body, &mut file).map(|_| ()))
+                .and_then(|_| std::fs::rename(&tmp_path, &local_path))
+            {
+                Ok(()) => {
+                    downloaded += 1;
+                    info!("媒体下载: {} -> {}", remote_path, local_path.display());
+                }
+                Err(e) => {
+                    let _ = std::fs::remove_file(&tmp_path);
+                    debug!("媒体下载写入失败 {}: {}", local_path.display(), e);
+                }
             }
         }
     }
@@ -778,8 +823,11 @@ pub fn cleanup_orphaned_remote_media(
         merged_map.iter().map(|e| e.hash.as_str()).collect();
 
     // 从远端文件名提取 hash（文件名格式: {hash}.{ext}）
+    fn remote_hash(filename: &str) -> Option<&str> {
+        filename.rsplit_once('.').map(|(hash, _)| hash)
+    }
     let orphan_files: Vec<&String> = remote_files.iter().filter(|f| {
-        if let Some(hash) = f.rsplit('.').last() {
+        if let Some(hash) = remote_hash(f) {
             !referenced_hashes.contains(hash)
         } else {
             false
@@ -814,7 +862,7 @@ pub fn cleanup_orphaned_remote_media(
 
     // 同步更新 media_map.json（移除已删除条目的引用）
     let orphan_hashes: std::collections::HashSet<&str> = orphan_files.iter().filter_map(|f| {
-        f.rsplit('.').last()
+        remote_hash(f)
     }).collect();
     if !orphan_hashes.is_empty() {
         if let Ok(mut map) = download_media_map(config) {
@@ -894,8 +942,9 @@ fn load_config_and_options(db: &crate::database::Database) -> Option<(WebDavConf
 
     let proxy_mode = repo.get("webdav_proxy_mode").ok().flatten().unwrap_or_else(|| "system".to_string());
     let proxy_url = repo.get("webdav_proxy_url").ok().flatten().unwrap_or_default();
+    let accept_invalid_certs = get_bool("webdav_accept_invalid_certs", false);
 
-    Some((WebDavConfig { url, username, password, remote_dir, proxy_mode, proxy_url }, options))
+    Some((WebDavConfig { url, username, password, remote_dir, proxy_mode, proxy_url, accept_invalid_certs }, options))
 }
 
 /// 用于追踪媒体同步是否正在进行
@@ -1145,9 +1194,28 @@ pub fn download_sync(config: &WebDavConfig, filename: &str) -> Result<Option<Vec
     let status = resp.status().as_u16();
     match status {
         200..=299 => {
-            let bytes = resp.bytes().map_err(|e| format!("读取响应失败: {}", e))?;
-            info!("WebDAV 下载成功: {} bytes", bytes.len());
-            Ok(Some(bytes.to_vec()))
+            if let Some(len) = resp.content_length() {
+                if len > MAX_SYNC_DOWNLOAD_BYTES {
+                    return Err(format!(
+                        "同步文件过大: {} bytes，超过 {} bytes 上限",
+                        len, MAX_SYNC_DOWNLOAD_BYTES
+                    ));
+                }
+            }
+
+            let mut limited = resp.take(MAX_SYNC_DOWNLOAD_BYTES + 1);
+            let mut data = Vec::new();
+            limited
+                .read_to_end(&mut data)
+                .map_err(|e| format!("读取响应失败: {}", e))?;
+            if data.len() as u64 > MAX_SYNC_DOWNLOAD_BYTES {
+                return Err(format!(
+                    "同步文件过大: 超过 {} bytes 上限",
+                    MAX_SYNC_DOWNLOAD_BYTES
+                ));
+            }
+            info!("WebDAV 下载成功: {} bytes", data.len());
+            Ok(Some(data))
         }
         404 => {
             info!("远端无同步文件");
