+++ src/components/TagsView.tsx
@@ -24,39 +24,12 @@ import {
   ReOrderDotsVertical16Regular,
   Tag16Regular,
   TagMultiple16Regular,
-  Document16Regular,
-  Folder16Regular,
-  Warning16Regular,
-  Video16Regular,
   TextSortAscending16Regular,
   Clock16Regular,
   ReOrderDotsVertical16Filled,
 } from "@fluentui/react-icons";
-import { invoke, convertFileSrc } from "@tauri-apps/api/core";
-import { emitTo } from "@tauri-apps/api/event";
-import { ImagePreview, getPreviewBounds } from "@/components/CardContentRenderers";
-import {
-  acquireTextPreviewLease,
-  revokeTextPreviewLease,
-  isTextPreviewLeaseCurrent,
-  isTextPreviewWanted,
-  textPreviewCleanupCallbacks,
-  ensureWindowHiddenListener,
-} from "@/components/ClipboardItemCard";
-import { HighlightText } from "@/components/HighlightText";
-import {
-  type ClipboardItemDetail,
-  sampleTextPreview,
-  getCachedTextPreviewContent,
-  setCachedTextPreviewContent,
-  TEXT_PREVIEW_MIN_W,
-  TEXT_PREVIEW_MAX_W,
-  TEXT_PREVIEW_MIN_H,
-  TEXT_PREVIEW_MAX_H,
-  TEXT_PREVIEW_CHAR_WIDTH,
-  TEXT_PREVIEW_HORIZONTAL_PADDING,
-  TEXT_PREVIEW_MIN_CHARS_PER_LINE,
-} from "@/components/text-preview";
+import { invoke } from "@tauri-apps/api/core";
+import { SortableTagItemRow, TagItemRow } from "@/components/tags/TagItemRows";
 import { Button } from "@/components/ui/button";
 import {
   Dialog,
@@ -73,7 +46,6 @@ import {
   TooltipTrigger,
 } from "@/components/ui/tooltip";
 import { focusWindowImmediately } from "@/hooks/useInputFocus";
-import { formatTime, contentTypeConfig, formatSize, getFileNameFromPath, parseFilePaths } from "@/lib/format";
 import { logError } from "@/lib/logger";
 import { cn } from "@/lib/utils";
 import { useClipboardStore, type ClipboardItem } from "@/stores/clipboard";
@@ -652,510 +624,6 @@ export function TagsView() {
           </DialogFooter>
         </DialogContent>
       </Dialog>
-    </div>
-  );
-}
-
-// ============ 文本悬浮预览 Hook ============
-
-function useTextHoverPreview(item: ClipboardItem) {
-  const textPreviewEnabled = useUISettings((s) => s.textPreviewEnabled);
-  const hoverPreviewDelay = useUISettings((s) => s.hoverPreviewDelay);
-  const previewPosition = useUISettings((s) => s.previewPosition);
-  const sharpCorners = useUISettings((s) => s.sharpCorners);
-
-  const isTextLike =
-    item.content_type === "text" || item.content_type === "html" || item.content_type === "rtf";
-
-  const anchorRef = useRef<HTMLDivElement>(null);
-  const timerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-  const hoveringRef = useRef(false);
-  const reqIdRef = useRef(0);
-  const leaseRef = useRef<number | null>(null);
-  const visibleRef = useRef(false);
-  const scrollRafRef = useRef<number | null>(null);
-  const scrollDeltaRef = useRef(0);
-
-  const clearTimer = useCallback(() => {
-    if (timerRef.current) {
-      clearTimeout(timerRef.current);
-      timerRef.current = null;
-    }
-  }, []);
-
-  const hidePreview = useCallback(() => {
-    reqIdRef.current += 1;
-    const closingLease = leaseRef.current;
-    if (closingLease !== null) {
-      revokeTextPreviewLease(closingLease);
-      leaseRef.current = null;
-    }
-    clearTimer();
-    hoveringRef.current = false;
-    if (scrollRafRef.current !== null) {
-      cancelAnimationFrame(scrollRafRef.current);
-      scrollRafRef.current = null;
-    }
-    scrollDeltaRef.current = 0;
-    if (closingLease !== null) {
-      visibleRef.current = false;
-      invoke("hide_text_preview", { token: closingLease }).catch((error) => {
-        logError("Failed to hide text preview:", error);
-      });
-    } else if (visibleRef.current) {
-      visibleRef.current = false;
-      invoke("hide_text_preview").catch((error) => {
-        logError("Failed to hide text preview:", error);
-      });
-    }
-  }, [clearTimer]);
-
-  const resolveContent = useCallback(async (): Promise<string> => {
-    const inlineText = item.text_content || item.preview || "";
-    if (!isTextLike) return "";
-    if (item.text_content) return item.text_content;
-    const cached = getCachedTextPreviewContent(item.id);
-    if (cached) return cached;
-    try {
-      const detail = await invoke<ClipboardItemDetail | null>("get_clipboard_item", { id: item.id });
-      const resolved = detail?.text_content || detail?.preview || inlineText;
-      if (resolved) setCachedTextPreviewContent(item.id, resolved);
-      return resolved;
-    } catch (error) {
-      logError("Failed to load full text content for preview:", error);
-      return inlineText;
-    }
-  }, [isTextLike, item.id, item.preview, item.text_content]);
-
-  const showPreview = useCallback(async (reqId: number, lease: number) => {
-    if (!textPreviewEnabled || !isTextLike || !anchorRef.current) return;
-    if (!hoveringRef.current || reqId !== reqIdRef.current || !isTextPreviewLeaseCurrent(lease)) return;
-    const textContent = await resolveContent();
-    if (!textContent) return;
-    if (!hoveringRef.current || reqId !== reqIdRef.current || !isTextPreviewLeaseCurrent(lease)) return;
-
-    const bounds = await getPreviewBounds(previewPosition, anchorRef.current);
-    if (!hoveringRef.current || reqId !== reqIdRef.current || !isTextPreviewLeaseCurrent(lease)) return;
-
-    const availableCssW = Math.max(260, Math.floor(bounds.maxW / bounds.scale));
-    const availableCssH = Math.max(140, Math.floor(bounds.maxH / bounds.scale));
-    const sampled = sampleTextPreview(textContent);
-    const desiredWidth = sampled.longestVisualCols * TEXT_PREVIEW_CHAR_WIDTH + TEXT_PREVIEW_HORIZONTAL_PADDING;
-    const windowCssW = Math.min(availableCssW, Math.min(TEXT_PREVIEW_MAX_W, Math.max(TEXT_PREVIEW_MIN_W, desiredWidth)));
-    const charsPerLine = Math.max(TEXT_PREVIEW_MIN_CHARS_PER_LINE, Math.floor((windowCssW - 30) / TEXT_PREVIEW_CHAR_WIDTH));
-    const sampledWrappedLines = sampled.lineColumns.reduce(
-      (sum, lineCols) => sum + Math.max(1, Math.ceil(lineCols / charsPerLine)), 0,
-    );
-    let estimatedLines = sampledWrappedLines;
-    if (sampled.truncated && sampled.processedCodeUnits < textContent.length) {
-      const remaining = textContent.length - sampled.processedCodeUnits;
-      const linesPerCodeUnit = sampledWrappedLines / Math.max(1, sampled.processedCodeUnits);
-      estimatedLines += Math.max(1, Math.ceil(remaining * linesPerCodeUnit));
-    }
-    const estimatedCssH = Math.min(TEXT_PREVIEW_MAX_H, Math.max(TEXT_PREVIEW_MIN_H, estimatedLines * 21 + 40));
-    const windowCssH = Math.min(availableCssH, estimatedCssH);
-    const winW = Math.max(1, Math.round(windowCssW * bounds.scale));
-    const winH = Math.max(1, Math.round(windowCssH * bounds.scale));
-    const winX = bounds.side === "left" ? bounds.anchorX - winW : bounds.anchorX;
-    const centeredY = Math.round(bounds.cardCenterY - winH / 2);
-    const winY = Math.max(bounds.monY, Math.min(centeredY, bounds.monBottom - winH));
-    const align = bounds.side === "left" ? "right" : "left";
-    const theme = document.documentElement.classList.contains("dark") ? "dark" : "light";
-
-    try {
-      invoke("hide_image_preview").catch((error) => logError("Failed to hide image preview:", error));
-      const uiState = useUISettings.getState();
-      await invoke("show_text_preview", {
-        text: textContent,
-        winX, winY, winWidth: winW, winHeight: winH,
-        align, theme, sharpCorners,
-        windowEffect: uiState.windowEffect,
-        fontFamily: uiState.previewFont || null,
-        fontSize: uiState.previewFontSize,
-        token: lease,
-      });
-      if (!hoveringRef.current || reqId !== reqIdRef.current || !isTextPreviewLeaseCurrent(lease)) {
-        visibleRef.current = false;
-        if (!isTextPreviewWanted()) {
-          invoke("hide_text_preview", { token: lease }).catch((error) => logError("Failed to hide text preview:", error));
-        }
-        return;
-      }
-      visibleRef.current = true;
-    } catch (error) {
-      visibleRef.current = false;
-      logError("Failed to show text preview:", error);
-    }
-  }, [textPreviewEnabled, isTextLike, previewPosition, resolveContent, sharpCorners]);
-
-  const handleMouseEnter = useCallback(() => {
-    if (!textPreviewEnabled || !isTextLike) return;
-    hoveringRef.current = true;
-    reqIdRef.current += 1;
-    const reqId = reqIdRef.current;
-    const lease = acquireTextPreviewLease();
-    leaseRef.current = lease;
-    clearTimer();
-    timerRef.current = setTimeout(() => {
-      void showPreview(reqId, lease);
-    }, hoverPreviewDelay);
-  }, [textPreviewEnabled, isTextLike, clearTimer, showPreview, hoverPreviewDelay]);
-
-  const handleMouseLeave = useCallback(() => {
-    hidePreview();
-  }, [hidePreview]);
-
-  const handleWheel = useCallback((e: React.WheelEvent<HTMLDivElement>) => {
-    if (!e.ctrlKey || !visibleRef.current) return;
-    e.preventDefault();
-    e.stopPropagation();
-    scrollDeltaRef.current += e.deltaY;
-    if (scrollRafRef.current === null) {
-      scrollRafRef.current = requestAnimationFrame(() => {
-        scrollRafRef.current = null;
-        const deltaY = scrollDeltaRef.current;
-        scrollDeltaRef.current = 0;
-        if (deltaY === 0 || !visibleRef.current) return;
-        emitTo("text-preview", "text-preview-scroll", { deltaY }).catch((error) => {
-          visibleRef.current = false;
-          logError("Failed to emit text preview scroll:", error);
-        });
-      });
-    }
-  }, []);
-
-  useEffect(() => {
-    return () => { hidePreview(); };
-  }, [hidePreview]);
-
-  useEffect(() => {
-    if (!textPreviewEnabled || !isTextLike) hidePreview();
-  }, [textPreviewEnabled, isTextLike, hidePreview]);
-
-  useEffect(() => {
-    ensureWindowHiddenListener();
-    textPreviewCleanupCallbacks.add(hidePreview);
-    return () => { textPreviewCleanupCallbacks.delete(hidePreview); };
-  }, [hidePreview]);
-
-  return { anchorRef, handleMouseEnter, handleMouseLeave, handleWheel };
-}
-
-// ============ Sortable Tag Item Row (drag-and-drop) ============
-
-function SortableTagItemRow({
-  item,
-  timeFormat,
-  onPaste,
-  onRemove,
-  isLast,
-}: {
-  item: ClipboardItem;
-  timeFormat: "relative" | "absolute";
-  onPaste: () => void;
-  onRemove: () => void;
-  isLast: boolean;
-}) {
-  const {
-    attributes,
-    listeners,
-    setNodeRef,
-    transform,
-    transition,
-    isDragging,
-  } = useSortable({ id: item.id });
-
-  const style = useMemo(() => ({
-    transform: CSS.Transform.toString(transform),
-    transition,
-    opacity: isDragging ? 0.5 : undefined,
-  }), [transform, transition, isDragging]);
-
-  const typeConfig = contentTypeConfig[item.content_type] ?? contentTypeConfig.text;
-  const preview = item.preview ?? item.text_content ?? "";
-  const truncated = preview.length > 120 ? preview.slice(0, 120) + "…" : preview;
-  const filePaths = (item.content_type === "files" || item.content_type === "video") ? parseFilePaths(item.file_paths) : [];
-  const filesInvalid = (item.content_type === "files" || item.content_type === "video") && item.files_valid === false;
-  const isVideo = item.content_type === "video";
-  const logicalLabel = isVideo ? contentTypeConfig.video.label : typeConfig.label;
-  const { anchorRef: textRef, handleMouseEnter: onTextEnter, handleMouseLeave: onTextLeave, handleWheel: onTextWheel } = useTextHoverPreview(item);
-
-  return (
-    <div
-      ref={setNodeRef}
-      style={style}
-      className={cn(
-        "flex items-start gap-1.5 px-2 py-2 hover:bg-accent/40 cursor-default group/row transition-colors duration-100",
-        !isLast && "border-b border-border/50",
-      )}
-      onClick={onPaste}
-    >
-      <button
-        {...attributes}
-        {...listeners}
-        className="shrink-0 mt-0.5 w-5 h-5 flex items-center justify-center text-muted-foreground/40 cursor-grab active:cursor-grabbing touch-none"
-        onClick={(e) => e.stopPropagation()}
-      >
-        <ReOrderDotsVertical16Regular className="w-3.5 h-3.5" />
-      </button>
-      <div className="flex-1 min-w-0">
-        {item.content_type === "image" && item.image_path ? (
-          <>
-            <ImagePreview
-              src={convertFileSrc(item.image_path)}
-              alt="Preview"
-              onError={() => {}}
-              imagePath={item.image_path}
-            />
-            <div className="flex items-center gap-1.5 mt-1.5">
-              <span className="inline-flex items-center text-[10px] text-muted-foreground px-1 py-px rounded bg-muted/60">
-                {typeConfig.label}
-              </span>
-              <span className="text-[10px] text-muted-foreground/60">
-                {formatTime(item.created_at, timeFormat)}
-              </span>
-              <span className="ml-auto text-[10px] text-muted-foreground/50">
-                {formatSize(item.byte_size)}
-              </span>
-            </div>
-          </>
-        ) : (item.content_type === "files" || item.content_type === "video") ? (
-          <>
-            <div className="flex items-center gap-2">
-              <div className={cn(
-                "shrink-0 w-8 h-8 rounded-md flex items-center justify-center",
-                filesInvalid ? "bg-red-50 dark:bg-red-950" : isVideo ? "bg-purple-50 dark:bg-purple-950" : "bg-blue-50 dark:bg-blue-950",
-              )}>
-                {filesInvalid ? (
-                  <Warning16Regular className="w-4 h-4 text-red-500" />
-                ) : isVideo ? (
-                  filePaths.length > 1 ? (
-                    <Folder16Regular className="w-4 h-4 text-purple-500" />
-                  ) : (
-                    <Video16Regular className="w-4 h-4 text-purple-500" />
-                  )
-                ) : filePaths.length > 1 ? (
-                  <Folder16Regular className="w-4 h-4 text-blue-500" />
-                ) : (
-                  <Document16Regular className="w-4 h-4 text-blue-500" />
-                )}
-              </div>
-              <div className="flex-1 min-w-0">
-                <p className={cn(
-                  "text-xs font-medium truncate",
-                  filesInvalid ? "text-red-500 line-through" : "text-foreground/90",
-                )}>
-                  {filePaths.length > 1
-                    ? `${filePaths.length} 个${isVideo ? "视频" : "文件"}`
-                    : getFileNameFromPath(filePaths[0] || preview)}
-                  {filesInvalid && <span className="ml-1 text-[10px] font-normal">(已失效)</span>}
-                </p>
-                {filePaths.length > 1 ? (
-                  <p className={cn(
-                    "text-[10px] truncate mt-0.5",
-                    filesInvalid ? "text-red-400 line-through" : "text-muted-foreground/60",
-                  )}>
-                    {filePaths.map(p => getFileNameFromPath(p)).slice(0, 3).join(", ")}{filePaths.length > 3 ? "..." : ""}
-                  </p>
-                ) : (
-                  <p className={cn(
-                    "text-[10px] truncate mt-0.5",
-                    filesInvalid ? "text-red-400 line-through" : "text-muted-foreground/60",
-                  )}>
-                    {filePaths[0] || preview || ""}
-                  </p>
-                )}
-              </div>
-            </div>
-            <div className="flex items-center gap-1.5 mt-1">
-              <span className="inline-flex items-center text-[10px] text-muted-foreground px-1 py-px rounded bg-muted/60">
-                {logicalLabel}
-              </span>
-              <span className="text-[10px] text-muted-foreground/60">
-                {formatTime(item.created_at, timeFormat)}
-              </span>
-              <span className="ml-auto text-[10px] text-muted-foreground/50">
-                {formatSize(item.byte_size)}
-              </span>
-            </div>
-          </>
-        ) : (
-          <div
-            ref={textRef}
-            onMouseEnter={onTextEnter}
-            onMouseLeave={onTextLeave}
-            onWheel={onTextWheel}
-          >
-            <p className="text-xs leading-relaxed line-clamp-2 break-all text-foreground/90"><HighlightText text={truncated || "(空)"} /></p>
-            <div className="flex items-center gap-1.5 mt-1">
-              <span className="inline-flex items-center text-[10px] text-muted-foreground px-1 py-px rounded bg-muted/60">
-                {typeConfig.label}
-              </span>
-              <span className="text-[10px] text-muted-foreground/60">
-                {formatTime(item.created_at, timeFormat)}
-              </span>
-            </div>
-          </div>
-        )}
-      </div>
-      <Tooltip>
-        <TooltipTrigger asChild>
-          <button
-            onClick={(e) => { e.stopPropagation(); onRemove(); }}
-            className="shrink-0 mt-0.5 w-5 h-5 flex items-center justify-center rounded-md text-muted-foreground/50 hover:text-destructive hover:bg-destructive/10 opacity-0 group-hover/row:opacity-100 transition-all duration-150"
-          >
-            <Dismiss16Regular className="w-3 h-3" />
-          </button>
-        </TooltipTrigger>
-        <TooltipContent>移除标签</TooltipContent>
-      </Tooltip>
-    </div>
-  );
-}
-
-// ============ Tag Item Row ============
-
-function TagItemRow({
-  item,
-  timeFormat,
-  onPaste,
-  onRemove,
-  isLast,
-}: {
-  item: ClipboardItem;
-  timeFormat: "relative" | "absolute";
-  onPaste: () => void;
-  onRemove: () => void;
-  isLast: boolean;
-}) {
-  const typeConfig = contentTypeConfig[item.content_type] ?? contentTypeConfig.text;
-  const preview = item.preview ?? item.text_content ?? "";
-  const truncated = preview.length > 120 ? preview.slice(0, 120) + "…" : preview;
-  const filePaths = (item.content_type === "files" || item.content_type === "video") ? parseFilePaths(item.file_paths) : [];
-  const filesInvalid = (item.content_type === "files" || item.content_type === "video") && item.files_valid === false;
-  const isVideo = item.content_type === "video";
-  const logicalLabel = isVideo ? contentTypeConfig.video.label : typeConfig.label;
-  const { anchorRef: textRef, handleMouseEnter: onTextEnter, handleMouseLeave: onTextLeave, handleWheel: onTextWheel } = useTextHoverPreview(item);
-
-  return (
-    <div
-      className={cn(
-        "flex items-start gap-2.5 px-3 py-2 hover:bg-accent/40 cursor-default group/row transition-colors duration-100",
-        !isLast && "border-b border-border/50",
-      )}
-      onClick={onPaste}
-    >
-      <div className="flex-1 min-w-0">
-        {item.content_type === "image" && item.image_path ? (
-          <>
-            <ImagePreview
-              src={convertFileSrc(item.image_path)}
-              alt="Preview"
-              onError={() => {}}
-              imagePath={item.image_path}
-            />
-            <div className="flex items-center gap-1.5 mt-1.5">
-              <span className="inline-flex items-center text-[10px] text-muted-foreground px-1 py-px rounded bg-muted/60">
-                {typeConfig.label}
-              </span>
-              <span className="text-[10px] text-muted-foreground/60">
-                {formatTime(item.created_at, timeFormat)}
-              </span>
-              <span className="ml-auto text-[10px] text-muted-foreground/50">
-                {formatSize(item.byte_size)}
-              </span>
-            </div>
-          </>
-        ) : (item.content_type === "files" || item.content_type === "video") ? (
-          <>
-            <div className="flex items-center gap-2">
-              <div className={cn(
-                "shrink-0 w-8 h-8 rounded-md flex items-center justify-center",
-                filesInvalid ? "bg-red-50 dark:bg-red-950" : isVideo ? "bg-purple-50 dark:bg-purple-950" : "bg-blue-50 dark:bg-blue-950",
-              )}>
-                {filesInvalid ? (
-                  <Warning16Regular className="w-4 h-4 text-red-500" />
-                ) : isVideo ? (
-                  filePaths.length > 1 ? (
-                    <Folder16Regular className="w-4 h-4 text-purple-500" />
-                  ) : (
-                    <Video16Regular className="w-4 h-4 text-purple-500" />
-                  )
-                ) : filePaths.length > 1 ? (
-                  <Folder16Regular className="w-4 h-4 text-blue-500" />
-                ) : (
-                  <Document16Regular className="w-4 h-4 text-blue-500" />
-                )}
-              </div>
-              <div className="flex-1 min-w-0">
-                <p className={cn(
-                  "text-xs font-medium truncate",
-                  filesInvalid ? "text-red-500 line-through" : "text-foreground/90",
-                )}>
-                  {filePaths.length > 1
-                    ? `${filePaths.length} 个${isVideo ? "视频" : "文件"}`
-                    : getFileNameFromPath(filePaths[0] || preview)}
-                  {filesInvalid && <span className="ml-1 text-[10px] font-normal">(已失效)</span>}
-                </p>
-                {filePaths.length > 1 ? (
-                  <p className={cn(
-                    "text-[10px] truncate mt-0.5",
-                    filesInvalid ? "text-red-400 line-through" : "text-muted-foreground/60",
-                  )}>
-                    {filePaths.map(p => getFileNameFromPath(p)).slice(0, 3).join(", ")}{filePaths.length > 3 ? "..." : ""}
-                  </p>
-                ) : (
-                  <p className={cn(
-                    "text-[10px] truncate mt-0.5",
-                    filesInvalid ? "text-red-400 line-through" : "text-muted-foreground/60",
-                  )}>
-                    {filePaths[0] || preview || ""}
-                  </p>
-                )}
-              </div>
-            </div>
-            <div className="flex items-center gap-1.5 mt-1">
-              <span className="inline-flex items-center text-[10px] text-muted-foreground px-1 py-px rounded bg-muted/60">
-                {logicalLabel}
-              </span>
-              <span className="text-[10px] text-muted-foreground/60">
-                {formatTime(item.created_at, timeFormat)}
-              </span>
-              <span className="ml-auto text-[10px] text-muted-foreground/50">
-                {formatSize(item.byte_size)}
-              </span>
-            </div>
-          </>
-        ) : (
-          <div
-            ref={textRef}
-            onMouseEnter={onTextEnter}
-            onMouseLeave={onTextLeave}
-            onWheel={onTextWheel}
-          >
-            <p className="text-xs leading-relaxed line-clamp-2 break-all text-foreground/90"><HighlightText text={truncated || "(空)"} /></p>
-            <div className="flex items-center gap-1.5 mt-1">
-              <span className="inline-flex items-center text-[10px] text-muted-foreground px-1 py-px rounded bg-muted/60">
-                {typeConfig.label}
-              </span>
-              <span className="text-[10px] text-muted-foreground/60">
-                {formatTime(item.created_at, timeFormat)}
-              </span>
-            </div>
-          </div>
-        )}
-      </div>
-      <Tooltip>
-        <TooltipTrigger asChild>
-          <button
-            onClick={(e) => { e.stopPropagation(); onRemove(); }}
-            className="shrink-0 mt-0.5 w-5 h-5 flex items-center justify-center rounded-md text-muted-foreground/50 hover:text-destructive hover:bg-destructive/10 opacity-0 group-hover/row:opacity-100 transition-all duration-150"
-          >
-            <Dismiss16Regular className="w-3 h-3" />
-          </button>
-        </TooltipTrigger>
-        <TooltipContent>移除标签</TooltipContent>
-      </Tooltip>
     </div>
   );
 }
